+++ src/mx_del_intarr.c
@@ -4,7 +4,8 @@ void mx_del_intarr(int **arr, int num)
 {
     if (!arr)
         return;
-    for (int i = 0; i < num; i++)
+
+    for (size_t i = 0; (int)i < num; i++)
         free(arr[i]);
     free(arr);
 }
